cognistep_env/pyproject.toml
[project]
name = "cognistep_env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "huggingface-hub>=1.8.0",
    "openai>=2.30.0",
    "openenv-core>=0.2.3",
    "pydantic>=2.12.5",
    "uvicorn>=0.42.0",
]
[project.scripts]
server = "cognistep_env.server.app:main"

[tool.setuptools]
packages = ["cognistep_env", "cognistep_env.server"]
package-dir = { "cognistep_env" = ".", "cognistep_env.server" = "server" }

[tool.uv]
package = true
